update how we save grammar concept results to cms (#11596)
