+++ services/QuillLMS/client/app/bundles/Grammar/actions/responses.ts
@@ -6,6 +6,8 @@ import { ActionTypes } from './actionTypes';
 
 import { requestDelete, requestGet, requestPost, requestPut, } from '../../../modules/request/index';
 import objectWithSnakeKeysFromCamel from '../libs/objectWithSnakeKeysFromCamel';
+import convertConceptResultsArrayToHash from '../libs/convertConceptResultsArrayToHash';
+
 
 export function deleteStatus(questionId: string) {
   return { type: ActionTypes.DELETE_RESPONSE_STATUS, data: { questionId, }, };
@@ -26,6 +28,8 @@ export function submitResponse(content: Response, prid: string, isFirstAttempt: 
   rubyConvertedResponse.created_at = moment().format('x');
   rubyConvertedResponse.first_attempt_count = isFirstAttempt ? 1 : 0;
   rubyConvertedResponse.is_first_attempt = isFirstAttempt;
+  rubyConvertedResponse.concept_results = convertConceptResultsArrayToHash(rubyConvertedResponse.conceptResults || rubyConvertedResponse.concept_results)
+
   return (dispatch: Function) => {
     requestPost(
       `${process.env.QUILL_CMS}/responses/create_or_increment`,
