+++ services/QuillLMS/client/app/bundles/Grammar/libs/grading/rematching.ts
@@ -4,6 +4,7 @@ import { checkGrammarQuestion, ConceptResult } from 'quill-marking-logic';
 import { requestGet, requestPost, requestPut, } from '../../../../modules/request/index';
 import { hashToCollection } from '../../../Shared/index';
 import objectWithSnakeKeysFromCamel from '../objectWithSnakeKeysFromCamel';
+import convertConceptResultsArrayToHash from '../convertConceptResultsArrayToHash'
 
 interface Question {
   conceptID: string,
@@ -146,7 +147,7 @@ function updateRematchedResponse(response: any, newResponse: any) {
     parent_id: newResponse.response.parent_id,
     author: newResponse.response.author,
     feedback: newResponse.response.feedback,
-    concept_results: convertResponsesArrayToHash(conceptResults),
+    concept_results: convertConceptResultsArrayToHash(conceptResults),
   };
   return updateResponse(response.id, newVals);
 }
@@ -170,7 +171,7 @@ function determineDelta(response: any, newResponse: any) {
   const parentIDChanged = (newResponse.response.parent_id ? Number(newResponse.response.parent_id) : null) !== response.parent_id;
   const authorChanged = newResponse.response.author !== response.author;
   const feedbackChanged = newResponse.response.feedback !== response.feedback;
-  const conceptResultsChanged = !_.isEqual(convertResponsesArrayToHash(conceptResults), response.concept_results);
+  const conceptResultsChanged = !_.isEqual(convertConceptResultsArrayToHash(conceptResults), response.concept_results);
   const changed = parentIDChanged || authorChanged || feedbackChanged || conceptResultsChanged;
   if (changed) {
     if (unmatched) {
@@ -237,15 +238,4 @@ function formatGradedResponses(jsonString: string): {[key: string]: Response} {
     delete resp.concept_results;
   });
   return bodyToObj;
-}
-
-function convertResponsesArrayToHash(crArray: any) {
-  const crs = _.values(crArray);
-  const newHash: {[key:string]: Boolean} = {};
-  _.each(crs, (val) => {
-    if (val.conceptUID && val.conceptUID.length > 0) {
-      newHash[val.conceptUID] = val.correct;
-    }
-  });
-  return newHash;
 }
